[LCR-57] refactor: constants type 분리
