+++ src/constants/dashboard/Colors.ts
@@ -1,0 +1,9 @@
+export const Colors = [
+  "#FF98C0", // 가장 작은 값
+  "#FFB4D1",
+  "#D9D9FC",
+  "#D3E5FF",
+  "#C5EFE8",
+  "#B3D2FF",
+  "#BCBCFA", // 가장 큰 값
+];
