+++ src/types/ReservationType.ts
@@ -1,0 +1,4 @@
+export type ReservationChartType = {
+  day: string;
+  value: number;
+};
